requirements.txt
langgraph>=0.3.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-xai>=0.2.0
langchain-core>=0.3.0
mem0ai>=0.1.0
httpx>=0.27.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.34.0
slack-bolt>=1.21.0
pydantic>=2.0.0
